src/third_party/blink/web_tests/external/wpt/webrtc/RTCDTMFSender-insertDTMF.https-expected.txt

This is a testharness.js-based test.
PASS insertDTMF() should succeed if tones contains valid DTMF characters
PASS insertDTMF() should throw InvalidCharacterError if tones contains invalid DTMF characters
FAIL insertDTMF() should throw InvalidStateError if transceiver is stopped transceiver.stop is not a function
PASS insertDTMF() should throw InvalidStateError if transceiver.currentDirection is recvonly
PASS insertDTMF() should throw InvalidStateError if transceiver.currentDirection is inactive
PASS insertDTMF() should set toneBuffer to provided tones normalized, with old tones overridden
PASS insertDTMF() after remove and close should reject
Harness: the test ran to completion.
